feat(sqlite): add use sqlite function
